Add in various attempts to get a sparkline to render
Mostly uploading for the purpose of remembering what hasn't worked in future.
Sparkline package doesn't seem to work. Plotly also not working. Attempting to save as taglist/string and convert. Think issue is likely to be with the javascript side of datatables. Reactable would potentially work much better but is not supported in PowerBI service.